+++ mui/src/features/ColumnChart/types.ts
@@ -1,0 +1,11 @@
+
+export interface ColumnChartData {
+  value: number,
+  column_name: string
+}
+
+export interface ColumnData {
+  value: number,
+  column_name: string,
+  height: number
+}